+++ dist-extension/service-worker-loader.js
@@ -1,1 +1,1 @@
-import './assets/index.ts-bhG1hcWB.js';
+import './assets/index.ts-Pzw2CLnn.js';